make components reusable
