+++ frontend/package.json
@@ -28,7 +28,7 @@
     "@mui/icons-material": "^5.11.11",
     "@mui/material": "^5.11.12",
     "@reduxjs/toolkit": "^1.9.3",
-    "codeforlife": "github:ocadotechnology/codeforlife-package-javascript#v1.14.1",
+    "codeforlife": "github:ocadotechnology/codeforlife-package-javascript#v1.14.2",
     "country-list": "^2.3.0",
     "formik": "^2.2.9",
     "react": "^18.2.0",
