+++ frontend/src/components/form/SchoolCountryField.tsx
@@ -1,0 +1,22 @@
+import React from 'react';
+import { getNames } from 'country-list';
+
+import {
+  AutocompleteField
+} from 'codeforlife/lib/esm/components/form';
+
+const SchoolCountryField: React.FC = () => {
+  return (
+    <AutocompleteField
+      options={getNames()}
+      textFieldProps={{
+        required: true,
+        name: 'country',
+        helperText: 'Country',
+        placeholder: 'Country'
+      }}
+    />
+  );
+};
+
+export default SchoolCountryField;
