+++ frontend/src/pages/teacherOnboarding/SchoolForm.tsx
@@ -1,24 +1,21 @@
 import React from 'react';
 import {
-  Typography,
-  InputAdornment
+  Typography
 } from '@mui/material';
-import {
-  Business as BusinessIcon
-} from '@mui/icons-material';
-import { getNames } from 'country-list';
 
 import {
   Form,
-  TextField,
-  AutocompleteField,
   SubmitButton
 } from 'codeforlife/lib/esm/components/form';
 
+import SchoolNameField from '../../components/form/SchoolNameField';
+import SchoolPostcodeField from '../../components/form/SchoolPostcodeField';
+import SchoolCountryField from '../../components/form/SchoolCountryField';
+
 const SchoolForm: React.FC<{
-  nextStep: () => void
+  onSubmit: () => void
 }> = ({
-  nextStep
+  onSubmit
 }) => {
     interface Values {
       school: string;
@@ -42,47 +39,15 @@ const SchoolForm: React.FC<{
           // TODO: call backend
           console.log(values);
           setSubmitting(false);
-          nextStep();
+          onSubmit();
         }}
         stackProps={{
           width: { xs: '100%', md: '50%' }
         }}
       >
-        <TextField
-          required
-          name='school'
-          helperText='Name of school or club'
-          placeholder='Name of school or club'
-          InputProps={{
-            endAdornment: (
-              <InputAdornment position='end'>
-                <BusinessIcon />
-              </InputAdornment>
-            )
-          }}
-        />
-        <TextField
-          required
-          name='postcode'
-          helperText='Postcode / Zipcode'
-          placeholder='Postcode / Zipcode'
-          InputProps={{
-            endAdornment: (
-              <InputAdornment position='end'>
-                <BusinessIcon />
-              </InputAdornment>
-            )
-          }}
-        />
-        <AutocompleteField
-          options={getNames()}
-          textFieldProps={{
-            required: true,
-            name: 'country',
-            helperText: 'Country',
-            placeholder: 'Country'
-          }}
-        />
+        <SchoolNameField />
+        <SchoolPostcodeField />
+        <SchoolCountryField />
         <SubmitButton>
           Create school or club
         </SubmitButton>
